Change info weight to 0.1, add seed data for user1, default username
- info interaction weight: 1 → 0.1
- interactions.weight column type: INTEGER → REAL
- frontend default username set to user1
- backend/seed.py: inserts Foundation, Dracula, Dune, The Mysterious Affair
  at Styles with reading_list + info interactions for user1

Run seed: cd backend && python seed.py

+++ backend/database.py
@@ -34,7 +34,7 @@ def init_db():
                 username TEXT NOT NULL,
                 book_id TEXT NOT NULL,
                 action TEXT NOT NULL,
-                weight INTEGER NOT NULL,
+                weight REAL NOT NULL,
                 timestamp TEXT DEFAULT (datetime('now'))
             );
